css: static font-size, less side margins
Because `max-width` is set in pixels, and `font-size` is relative to the display width, as the screen gets wider, the *less* characters fit on a line. This gets especially absurd on ultrawide displays.

AFAIK the default font size should always be readable, so I just removed the explicit `font-size` completely, which fixes the issue. `max-width` could also be changed to be measured in `ch`, to make the characters/line consistent across fonts, but I wanted to change as little as possible.

I also removed the `ul` padding. It doesn't really do anything on large screens, and I think the additional space on narrow screens is worth it.